+++ src/framing/test_framing.rs
@@ -34,7 +34,7 @@ fn codec() {
         orig.signature = signature_input.sign(&credential_bundle);
 
         let enc = orig.encode_detached().unwrap();
-        let copy = MLSPlaintext::from_bytes(&enc).unwrap();
+        let copy = MLSPlaintext::decode(&mut Cursor::new(&enc)).unwrap();
         assert_eq!(orig, copy);
         assert!(!orig.is_handshake_message());
     }
